+++ esp_video/examples/common_components/example_video_common/include/boards/esp32-s31-korvo/example_video_common_board.h
@@ -64,7 +64,11 @@ extern "C" {
  */
 #define EXAMPLE_SPI_CAM_0_CS_PIN                        56
 #define EXAMPLE_SPI_CAM_0_SCLK_PIN                      54
+#ifdef CONFIG_CAMERA_BF3901
 #define EXAMPLE_SPI_CAM_0_DATA0_IO_PIN                  53
+#else
+#define EXAMPLE_SPI_CAM_0_DATA0_IO_PIN                  46
+#endif
 #define EXAMPLE_SPI_CAM_0_DATA1_IO_PIN                  -1
 
 /**